+++ crates/playwright/src/protocol/mod.rs
@@ -50,7 +50,7 @@ pub use api_request_context::APIRequestContext;
 pub use browser::Browser;
 pub use browser_context::{
     BrowserContext, BrowserContextOptions, BrowserContextOptionsBuilder, Cookie, Geolocation,
-    LocalStorageItem, Origin, StorageState, Viewport,
+    LocalStorageItem, Origin, RecordHar, RecordVideo, StorageState, Viewport,
 };
 pub use browser_type::BrowserType;
 pub use click::{ClickOptions, KeyboardModifier, MouseButton, Position};